@@ -25,19 +25,18 @@
         }
     ],
     "require": {
-        "php": "~7.4.0 || ~8.0.0 || ~8.1.0 || ~8.2.0",
-        "typo3/cms-core": "^11.5",
-        "typo3/cms-frontend": "^11.5",
-        "matomo/device-detector": "^6.1",
-        "symfony/http-foundation": "^5.4",
-        "psr/http-message": "^1.0"
+        "php": "~8.2.0 || ~8.3.0 || ~8.4.0",
+        "typo3/cms-core": "^13.4",
+        "typo3/cms-frontend": "^13.4",
+        "matomo/device-detector": "^6.4",
+        "symfony/http-foundation": "^7.2",
+        "psr/http-message": "^2.0"
     },
     "require-dev": {
-        "phpstan/phpstan": "^1.10",
-        "typo3/testing-framework": "^6.16",
-        "phpunit/phpunit": "^9.6",
-        "friendsofphp/php-cs-fixer": "^3.14",
-        "cweagans/composer-patches": "^1.7"
+        "phpstan/phpstan": "^2.1",
+        "typo3/testing-framework": "^9.2",
+        "phpunit/phpunit": "^11.5 || ^12.1",
+        "friendsofphp/php-cs-fixer": "^3.75"
     },
     "autoload": {
         "psr-4": {
@@ -54,11 +53,6 @@
             "cms-package-dir": "{$vendor-dir}/typo3/cms",
             "extension-key": "abtest",
             "web-dir": ".Build/web"
-        },
-        "patches": {
-            "typo3/testing-framework": {
-                "Allow to test requests with cookies": "patches/testing-framework-cookies.patch"
-            }
         }
     },
     "scripts": {
@@ -68,10 +62,10 @@
         ]
     },
     "config": {
+        "lock": false,
         "allow-plugins": {
             "typo3/class-alias-loader": true,
-            "typo3/cms-composer-installers": true,
-            "cweagans/composer-patches": true
+            "typo3/cms-composer-installers": true
         }
     }
 }
